Fix order issue when using std::set with UHDM models
The order of elements in std::set is unpredictable because it depends on
the pointer address which isn't guaranteed to be monotonically
increasing. This inconsistency ripples down to differences in order in
whcih objects are processed and thus affecting the reproducability of
output logs. The problem is even more glaring with compiler differences
on windows vs. linux.

To fix the issue, contrain the set on the model's id which are
guaranteed to be unique and monotonically increasing. This ensures that
the elements are still unique in the container and are in also in a
fixed, reproducible order.